requirements.txt
cachetools==4.2.4
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cssselect2==0.4.1
defusedxml==0.7.1
Flask==2.0.2
google-api-client==3.14.159265359.post1
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==3.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tinycss2==1.1.0
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
